Updated expected output file for functional tests
NOTE: I did not check the changes since there has been a huge change to our parser and test files. So just assume that everything is corret now and deal with bug when they appear. In the future, whenever tests fail and the expected output files are regenerated, it should of course be checked that the changes to the exptected output files make sense in repstect to the changes to the library.