Fix test collection issue because pytest replaces stdin
The issue is, that since pytest replaces stdin, the collection of the stdin buffer in utils fails. This has been fixed by implementing a context manager that temporarily replaces stdin with a Mock. This context manager should then be used to change the context of import of the pytest modules.